refactor: move FinalEvaluation signature
